qwen3-embedder-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qwen3-embedder"
version = "0.1.0"
description = "Multi-backend embedding service using Qwen3-Embedding models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Brennan Conley"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.24.0",
    "transformers>=4.40.0",
    "tokenizers>=0.15.0",
    "huggingface-hub>=0.20.0",
    "PyYAML>=6.0",
    "structlog>=24.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
cuda = [
    "torch>=2.1.0",
    "accelerate>=0.25.0",
    "flash-attn>=2.5.0; platform_system == 'Linux'",
]
cpu = [
    "torch>=2.1.0",
]
vllm = [
    "vllm>=0.4.0",
]
mlx = [
    "mlx>=0.10.0",
    "mlx-lm>=0.10.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
all = [
    "qwen3-embedder[cuda,vllm,mlx,dev]",
]

[project.scripts]
qwen3-embedder = "qwen3_embedder.main:main"

[project.urls]
Homepage = "https://github.com/your-org/qwen3-embedder-multi"
Repository = "https://github.com/your-org/qwen3-embedder-multi"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"
